refactory: 프로필 상세 페이지, 프로필 수정, 게시글 작성, 프로필 API 리팩토링
