+++ moamoa/src/Pages/Profile/YourProfile.jsx
@@ -2,7 +2,7 @@
   설명: 사용자 accountname의 프로필 페이지(남의 페이지)
   작성자: 이해지
   최초 작성 날짜: 2023.10.23
-  마지막 수정 날까: 2023.11.05
+  마지막 수정 날까: 2023.12.08
 */
 
 import React, { useState, useEffect } from 'react';
@@ -13,7 +13,7 @@ import { useLocation } from 'react-router-dom';
 
 import ProfileDetail from '../../Components/Profile/ProfileDetail';
 
-import FollowButton from '../../Components/Common/FollowButton';
+import FollowButton from '../../Components/Profile/FollowButton';
 import ProfileDetailPost from '../../Components/Profile/ProfileDetailPost';
 import ProfileDetailProduct from '../../Components/Profile/ProfileDetailProduct';
 import styled from 'styled-components';
@@ -26,7 +26,7 @@ import Footer from '../../Components/Common/Footer';
 import { Container } from '../../Components/Common/Container';
 import HeaderKebab from '../../Components/Common/HeaderKebab';
 
-import GetYourinfoAPI from '../../API/Profile/GetYourinfoAPI';
+import { getYourProfileData } from '../../API/Profile/ProfileAPI';
 
 // 프로필보기
 function YourProfile() {
@@ -48,7 +48,7 @@ function YourProfile() {
     setIsLoading(true); // API 호출 전에 로딩 상태를 true로 설정
     try {
       const infoUrl = location.pathname;
-      const res = await GetYourinfoAPI(infoUrl, token);
+      const res = await getYourProfileData(infoUrl);
 
       setProfileImg(res.profile['image']);
       setProfileAccountname(res.profile['accountname']);
@@ -105,11 +105,10 @@ function YourProfile() {
         <HiddenH1>남의 프로필</HiddenH1>
         <section>
           <ProfileTop>
-            <ProfileDetail userInfoData={userInfoData} token={token} />
+            <ProfileDetail userInfoData={userInfoData} />
             <Btns>
               <FollowButton
                 userAccount={userInfoData.profileAccountname}
-                token={token}
                 isFollow={!isLoading && userInfoData.isFollow}
                 reFetchInfo={UserInfo}
               />
@@ -131,8 +130,10 @@ function YourProfile() {
 
           {/* 라우터에 연결되면 채팅방으로 연결 */}
         </section>
-        {userType === 'organization' ? <ProfileDetailProduct /> : null}
-        <ProfileDetailPost />
+        {userType === 'organization' ? (
+          <ProfileDetailProduct userInfoData={userInfoData} reFetchInfo={UserInfo} />
+        ) : null}
+        <ProfileDetailPost accountName={profileAccountname} />
         <Footer />
       </section>
     </Container>
